Separate tests by entity
- Have different tests (in independent files) for each rule
- Create constants that can be reused across files
- Remove old file with all tests in one place.
