+++ cypress/integration/button_spec.js
@@ -1,0 +1,58 @@
+import { 
+  ERROR_BORDER_COLOR
+} from "../support/constants";
+
+describe("<button>", () => {
+  before(() => {
+    cy.visit("/test/index.html");
+  });
+
+  it('should not have buttons with interactive content inside', () => {
+    const noneElements = [
+      "audio[controls]", "embed", "iframe", "img[usemap]", 'input:not([type="hidden"])',
+      "object[usemap]", "select", "textarea", "video[controls]"
+    ];
+    // removed "button" from afterElements because browser was correcting the issue
+    const afterElements = ["details", "label"];
+    const beforeElements = [];
+
+    noneElements.forEach(el => {
+      cy.get(`button ${el}`)
+        .each(element => {
+          cy.get(element)
+            .should('have.css', 'border-color', ERROR_BORDER_COLOR)
+        });
+    });
+
+    afterElements.forEach(el => {
+      cy.get(`button ${el}`)
+        .each(element => {
+          if (element) {
+            cy.get(element)
+              .after('content')
+              .should('eq', `ERROR: Ensure that <${el}> is not a child of <button>.`);
+          }
+        });
+    });
+
+    beforeElements.forEach(el => {
+      cy.get(`button ${el}`)
+        .each(element => {
+          if (element) {
+            cy.get(element)
+              .before('content')
+              .should('eq', `ERROR: Ensure that <${el}> is not a child of <button> as it is an invalid HTML.`);
+          }
+        });
+    })
+  });
+
+  it('should not have empty buttons', () => {
+    cy.get('button:not( [aria-label] ):not( [aria-labelledby] ):empty')
+      .each(element => {
+        cy.get(element)
+          .after("content")
+          .should('eq', "ERROR: Ensure that <button> has meaningful content or is labelled appropriately.")
+      });
+  });
+});
