+++ cypress/support/constants.js
@@ -1,0 +1,3 @@
+// this is how Cypress reads the CSS. It will need to be changed if colors change!
+export const ERROR_BORDER_COLOR = 'rgb(255, 0, 0)';
+export const ERROR_WARNING_COLOR = 'rgb(255, 255, 102)';
